split fstat caching
